pyproject.toml
[tool.poetry]
name = "generative-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
python-dotenv = "^1.0.1"
# google-generativeai = "^0.4.0"  # this is not supported by langchain-google-genai
google-generativeai = ">=0.3.1,<0.4.0"
crewai = "^0.19.0"
duckduckgo-search = "^5.0"
huggingface-hub = "^0.21.4"
langchain-google-genai = "^0.0.9"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
